Fixed Overview and Reload
