+++ source/metroid3/samus/sprite.py
@@ -9,8 +9,6 @@ class Sprite(SpriteParent):
 	def __init__(self, filename, manifest_dict, my_subpath):
 		super().__init__(filename, manifest_dict, my_subpath)
 
-		self.overview_scale_factor = 1    #Samus's sheet is BIG, so don't zoom in on the overview
-
 		self.overhead = False   #Samus is sideview, so only left/right direction buttons should show
 
 		#TODO: Make file select and ship be animations in the big list, or tie to the ship background
