+++ user_resources/metroid3/backgrounds/backgrounds.json
@@ -1,7 +1,8 @@
 [
 	{
 		"filename": "ship.png",
-		"title": "Samus' Ship"
+		"title": "Samus's Ship",
+		"snarky_comment": "Even though Samus's name ends in an 'S', because she is only one person, there is still an 's' after the apostrophe."
 	},
 	{
 		"filename": "zebes.png",
